Move taskbar and style to App.vue
