+++ src/app/views/Index.vue
@@ -1,14 +1,9 @@
 <template>
-  <div class="app-desktop" :class="settingsStore.themeName" :style="{backgroundImage, backgroundColor}">
-    <component :is="window.component" v-for="window in openedWindows" :key="window.name" :name="window.name" />
-
-    <win-taskbar />
-    <win-status-bar />
-  </div>
+  <component :is="window.component" v-for="window in openedWindows" :key="window.name" :name="window.name" />
 </template>
 
 <script setup lang="ts">
-import { computed, onMounted } from 'vue'
+import { onMounted } from 'vue'
 import { useSettingsStore } from '@app/stores/settingsStore'
 import { useUserAuthStore } from '@app/stores/userAuthStore'
 import { useWindows } from '@app/composables/useWindows.ts'
@@ -16,10 +11,6 @@ import { useWindows } from '@app/composables/useWindows.ts'
 const settingsStore = useSettingsStore()
 const userAuthStore = useUserAuthStore()
 const { openWindow, openedWindows, WinType } = useWindows()
-
-// Appearance
-const backgroundImage = computed(() => settingsStore.backgroundSrc)
-const backgroundColor = computed(() => settingsStore.background.color)
 
 onMounted(() => {
   openWindow(WinType.LOADING)
